backend/requirements.txt
# Web framework
fastapi==0.111.1
uvicorn[standard]==0.23.2

# Database & ORM
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7

# Environment variables
python-dotenv==1.0.1

# Pydantic
pydantic==2.3.0

# Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
sentence-transformers==2.2.2

# Vector DB
faiss-cpu==1.7.4
langchain==0.1.156
langchain-community==0.0.33

# Optional async PostgreSQL
asyncpg==0.27.1
